Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/workers/baseurl/alpha/xhr-in-worker-expected.txt


FAIL xhr-worker assert_equals: expected "gamma\n" but got "beta\n"
